Delete invoice

+++ src/app/invoices/components/invoice-list/invoice-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { InvoiceService } from '../../services/invoice.service';
 import { Invoice } from '../../../shared/models/invoice';
+import { MatDialog, MatDialogRef } from '@angular/material';
+import { ConfirmDialogComponent } from '../../../shared/components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-invoice-list',
@@ -11,18 +13,24 @@ export class InvoiceListComponent implements OnInit {
   invoices: Invoice[];
   displayedColumns = ['date', 'number', 'to', 'from', 'operations'];
 
-  constructor(private invoiceService: InvoiceService) { }
+  constructor(private invoiceService: InvoiceService, public dialog: MatDialog) { }
 
   ngOnInit() {
     this.getItems();
   }
 
   delete(index: number): void {
-    this.invoiceService.remove(index).subscribe(
-      (result) => {
-        this.getItems();
+    const dialogRef = this.dialog.open(ConfirmDialogComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      console.log({result});
+      if (result === 'asdaqwe') {
+        this.invoiceService.remove(index).subscribe(
+          (data) => {
+            this.getItems();
+          }
+        );
       }
-    );
+    });
   }
 
   getItems() {
